Revert "Abstarcting the sanatization process into Sanatize class"
This reverts commit c2f2c368d1ae808d0ca9b9f3320e2a680d892c94.
