+++ app/Http/Controllers/CalculatorController.php
@@ -78,7 +78,7 @@ class CalculatorController extends Controller
     {
         $lettersCount = [];
         $lettersCount["sura_title"] = $this->fullSura->name;
-        $lettersCount["occurrences"] = $this->counter->countLettersInString($this->fullSura->suraString);
+        $lettersCount["occurrences"] = $counter->countLettersInString($this->fullSura->suraString);
 
         return $this->jsonResponse($lettersCount);
     }
