--- storage/decoded_suras/114الناس sura results
@@ -1,1 +1,0 @@
-{"name":"114الناس","numberOfVerses":6,"numberOfWords":20,"numberOfLetters":80,"wordOccurrences":{"قل":1,"اعوذ":1,"برب":1,"ملك":1,"اله":1,"شر":1,"الوسواس":1,"الخناس":1,"الذي":1,"يوسوس":1,"في":1,"صدور":1,"الجنة":1,"والناس":1,"من":2,"الناس":4},"wordIndex":{"قل":[1],"اعوذ":[2],"برب":[3],"الناس":[4,6,8,17],"ملك":[5],"اله":[7],"من":[9,18],"شر":[10],"الوسواس":[11],"الخناس":[12],"الذي":[13],"يوسوس":[14],"في":[15],"صدور":[16],"الجنة":[19],"والناس":[20]},"letterOccurrences":{"ا":18,"ب":2,"ج":1,"خ":1,"د":1,"ذ":2,"ر":3,"س":10,"ش":1,"ص":1,"ع":1,"ف":1,"ق":1,"ك":1,"ل":12,"م":3,"ن":9,"ه":1,"و":7,"ي":3},"LettersMapIndexes":{"ق":[1],"ل":[2,14,19,23,28,32,43,51,57,76,84,91]," ":[3,8,12,21,30,38,41,49,60,66,69,74,82,88],"ا":[4,13,16,22,25,27,31,34,42,47,50,54,56,75,78,83,90,93],"ع":[5],"و":[6,44,46,62,64,72,89],"ذ":[7,58],"ب":[9,11],"ر":[10,40,73],"ن":[15,24,33,37,53,77,81,86,92],"س":[17,26,35,45,48,55,63,65,79,94],"م":[18,36,80],"ك":[20],"ه":[29],"ش":[39],"خ":[52],"ي":[59,61,68],"ف":[67],"ص":[70],"د":[71],"ج":[85],"ة":[87]},"VerseIndex":{"1":"قل اعوذ برب الناس","2":"ملك الناس","3":"اله الناس","4":"من شر الوسواس الخناس","5":"الذي يوسوس في صدور الناس","6":"من الجنة والناس"}}
